requirements.txt
# --- Core Logic ---
pandas
gspread
google-auth
oauth2client
openpyxl
python-dotenv
openai

# --- Browser Automation ---
playwright

# --- UI / Dashboard ---
streamlit

# --- Utilities ---
colorlog

# --- Testing (Crucial for Dev) ---
pytest
pytest-mock